Move incentive stake operations to useBulkPosition Hook
