+++ apps/web/src/hooks/useBulkPosition.tsx
@@ -1,0 +1,280 @@
+import { useAccount } from "hooks/useAccount";
+import { useV3StakerContract } from "hooks/useV3StakerContract";
+import { useV3NFTPositionManagerContract } from "hooks/useContract";
+import { useCallback, useState, useEffect } from "react";
+import { IncentiveKey } from "hooks/usePosition";
+import { ProcessedIncentive } from "hooks/useIncentivesData";
+import { ethers } from "ethers";
+
+export const useBulkPosition = (tokenId: number, poolAddress: string, allIncentives: ProcessedIncentive[]) => {
+  const { address } = useAccount();
+  const v3StakerContract = useV3StakerContract();
+  const nftManagerPositionsContract = useV3NFTPositionManagerContract();
+  const [isBulkStaking, setIsBulkStaking] = useState(false);
+  const [isBulkUnstaking, setIsBulkUnstaking] = useState(false);
+  const [isBulkWithdrawing, setIsBulkWithdrawing] = useState(false);
+  const [isStaking, setIsStaking] = useState(false);
+  const [isUnstaking, setIsUnstaking] = useState(false);
+  const [isClaiming, setIsClaiming] = useState(false);
+  const [currentIncentiveId, setCurrentIncentiveId] = useState<string | null>(null);
+  const [hasRewards, setHasRewards] = useState(false);
+
+  useEffect(() => {
+    const checkRewards = async () => {
+      if (!v3StakerContract || !address) {
+        setHasRewards(false);
+        return;
+      }
+
+      try {
+        let totalRewards = 0;
+        for (const incentive of allIncentives) {
+          if (incentive.hasUserPositionInIncentive) {
+            const reward = await v3StakerContract.rewards(incentive.rewardToken.id, address);
+            const rewardAmount = ethers.utils.formatUnits(reward, 18);
+            totalRewards += Number(rewardAmount);
+          }
+        }
+        setHasRewards(totalRewards > 0);
+      } catch (error) {
+        console.error('Error checking rewards:', error);
+        setHasRewards(false);
+      }
+    };
+
+    checkRewards();
+  }, [v3StakerContract, address, allIncentives]);
+
+  const getIncentiveData = useCallback(async (incentiveId: string) => {
+    const incentive = allIncentives.find(inc => inc.id === incentiveId);
+    if (!incentive) {
+      return null
+    }
+    return incentive;
+  }, [allIncentives]);
+
+  const handleStake = useCallback(async (incentive: ProcessedIncentive) => {
+    if (!v3StakerContract || !address || !nftManagerPositionsContract) return;
+    setIsStaking(true);
+    setCurrentIncentiveId(incentive.id);
+
+    try {
+      const incentiveData = await getIncentiveData(incentive.id);
+      if (!incentiveData) {
+        throw new Error('Failed to fetch incentive data');
+      }
+
+      const incentiveKey: IncentiveKey = {
+        rewardToken: incentiveData.rewardToken.id,
+        pool: incentiveData.poolAddress,
+        startTime: incentiveData.startTime,
+        endTime: incentiveData.endTime,
+        vestingPeriod: parseInt(incentiveData.vestingPeriod),
+        refundee: incentiveData.refundee,
+      };
+
+      const stakeTx = await v3StakerContract.stakeToken(incentiveKey, tokenId);
+      await stakeTx.wait();
+    } catch (error) {
+      console.error('Error staking:', error);
+    } finally {
+      setIsStaking(false);
+      setCurrentIncentiveId(null);
+    }
+  }, [v3StakerContract, tokenId, address, getIncentiveData, nftManagerPositionsContract]);
+
+  const handleUnstake = useCallback(async (incentive: ProcessedIncentive) => {
+    if (!v3StakerContract || !address || !nftManagerPositionsContract) return;
+    setIsUnstaking(true);
+    setCurrentIncentiveId(incentive.id);
+
+    try {
+      const incentiveData = await getIncentiveData(incentive.id);
+      if (!incentiveData) {
+        throw new Error('Failed to fetch incentive data');
+      }
+
+      const incentiveKey: IncentiveKey = {
+        rewardToken: incentiveData.rewardToken.id,
+        pool: incentiveData.poolAddress,
+        startTime: incentiveData.startTime,
+        endTime: incentiveData.endTime,
+        vestingPeriod: parseInt(incentiveData.vestingPeriod),
+        refundee: incentiveData.refundee,
+      };
+
+      const unstakeTx = await v3StakerContract.unstakeToken(incentiveKey, tokenId);
+      await unstakeTx.wait();
+    } catch (error) {
+      console.error('Error unstaking:', error);
+    } finally {
+      setIsUnstaking(false);
+      setCurrentIncentiveId(null);
+    }
+  }, [v3StakerContract, tokenId, address, getIncentiveData, nftManagerPositionsContract]);
+
+  const handleClaim = useCallback(async (incentive: ProcessedIncentive) => {
+    if (!v3StakerContract || !address) return;
+    setIsClaiming(true);
+    setCurrentIncentiveId(incentive.id);
+
+    try {
+      const incentiveData = await getIncentiveData(incentive.id);
+      if (!incentiveData) {
+        throw new Error('Failed to fetch incentive data');
+      }
+
+      const reward = await v3StakerContract.rewards(
+        incentive.rewardToken.id,
+        address
+      );
+      const claimTx = await v3StakerContract.claimReward(
+        incentive.rewardToken.id,
+        address,
+        reward
+      );
+      const claimReceipt = await claimTx.wait();
+    } catch (error) {
+      console.error('Error claiming:', error);
+    } finally {
+      setIsClaiming(false);
+      setCurrentIncentiveId(null);
+    }
+  }, [v3StakerContract, address, getIncentiveData]);
+
+  const handleBulkStake = useCallback(async () => {
+    if (!v3StakerContract || !address || !nftManagerPositionsContract) return;
+    setIsBulkStaking(true);
+
+    try {
+      const incentivesToStake = allIncentives.filter(
+        (incentive) => !incentive.hasUserPositionInIncentive
+      );
+
+      if (incentivesToStake.length === 0) {
+        throw new Error('No incentives available to stake');
+      }
+
+      const incentiveKeys = await Promise.all(
+        incentivesToStake.map(async (incentive) => {
+            console.log('incentive.rewardToken.id,', incentive.rewardToken.id,)
+          const key = [
+            incentive.rewardToken.id,
+            incentive.poolAddress,
+            incentive.startTime,
+            incentive.endTime,
+            incentive.vestingPeriod,
+            incentive.refundee
+          ];
+          return key;
+        }))
+
+      const data = ethers.utils.defaultAbiCoder.encode(
+        ['tuple(address,address,uint256,uint256,uint256,address)[]'],
+        [incentiveKeys]
+      );
+
+      const transferTx = await nftManagerPositionsContract[
+        'safeTransferFrom(address,address,uint256,bytes)'
+      ](address, v3StakerContract.address, tokenId, data, {
+        gasLimit: 500000
+      });
+
+      const receipt = await transferTx.wait();
+      console.log('receipt', receipt)
+
+    } catch (error) {
+      console.error('Error in bulk staking:', error);
+      throw error;
+    } finally {
+      setIsBulkStaking(false);
+    }
+  }, [v3StakerContract, tokenId, address, allIncentives, nftManagerPositionsContract]);
+  
+
+  const handleBulkUnstake = useCallback(async () => {
+    if (!v3StakerContract || !address || !nftManagerPositionsContract) return;
+    setIsBulkUnstaking(true);
+
+    try {
+      const stakedIncentives = allIncentives.filter(
+        (incentive) => incentive.hasUserPositionInIncentive
+      );
+
+      if (stakedIncentives.length === 0) {
+        throw new Error('No staked incentives to unstake from');
+      }
+
+      const incentiveKeys = await Promise.all(
+        stakedIncentives.map(async (incentive) => {
+          const key = [
+            incentive.rewardToken.id,
+            incentive.poolAddress,
+            incentive.startTime,
+            incentive.endTime,
+            incentive.vestingPeriod,
+            incentive.refundee
+          ];
+          return key;
+        })
+      );
+
+      const unstakeCalls = incentiveKeys.map((key) => {
+        const callData = v3StakerContract.interface.encodeFunctionData('unstakeToken', [
+          key,
+          tokenId
+        ]);
+        return callData;
+      });
+
+      const unstakeTx = await v3StakerContract.multicall(unstakeCalls, {
+        gasLimit: 500000
+      });
+
+      const receipt = await unstakeTx.wait();
+
+    } catch (error) {
+      console.error('Error in bulk unstaking:', error);
+      throw error;
+    } finally {
+      setIsBulkUnstaking(false);
+    }
+  }, [v3StakerContract, tokenId, address, allIncentives, nftManagerPositionsContract]);
+  
+
+  const handleBulkWithdraw = useCallback(async () => {
+    if (!v3StakerContract || !address || !nftManagerPositionsContract) return;
+    setIsBulkWithdrawing(true);
+    try {
+      const withdrawTx = await v3StakerContract.withdrawToken(
+        tokenId,
+        address,
+        []
+      );
+      const withdrawReceipt = await withdrawTx.wait();
+    } catch (error) {
+      console.error('Error in bulk withdrawal:', error);
+      throw error;
+    } finally {
+      setIsBulkWithdrawing(false);
+    }
+  }, [v3StakerContract, tokenId, address, nftManagerPositionsContract]);
+
+  return {
+    isBulkStaking,
+    isBulkUnstaking,
+    isBulkWithdrawing,
+    isStaking,
+    isUnstaking,
+    isClaiming,
+    currentIncentiveId,
+    hasRewards,
+    getIncentiveData,
+    handleStake,
+    handleUnstake,
+    handleClaim,
+    handleBulkStake,
+    handleBulkUnstake,
+    handleBulkWithdraw
+  };
+};
